+++ ChangeLog.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] 
+### Added
+- error_t::message() function to retrieve details about exceptions
+- custom python exception type: ifm3dpy.Error
 
 ## 0.92.0 - 2021-10-22
 ### Added